Updated first login modal window elements
